+++ src/util/index.ts
@@ -20,3 +20,7 @@ export function sortThemesByType(
 		}
 	}
 }
+
+export function getThemeTypeLabel(themeType: ThemeType): "Dark" | "Light" {
+	return themeType === ThemeType.dark ? "Dark" : "Light";
+}