+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # Change Log
 
+## 1.2.0
+
+Add optional details in the picker - showing dark/light theme (`favouriteThemes.showExtraQuickPickDetails`)
+
+## 1.1.1
+
+Rework & clean up settings management
+
 ## 1.1.0
 
 ### Added
 
-- Sorting of pinned themes by recent usage
+- Sorting of pinned themes by recent usage (`favouriteThemes.sortPinnedByRecentUsage`)
 
 ## 1.0.0